+++ src/types/ray.rs
@@ -1,4 +1,4 @@
-use crate::types::Vec3;
+use ultraviolet::vec::Vec3;
 
 pub struct Ray {
     a: Vec3,
@@ -18,7 +18,7 @@ impl Ray {
         self.b
     }
     #[inline]
-    pub fn point_at_parameter(&self, t: f64) -> Vec3 {
+    pub fn point_at_parameter(&self, t: f32) -> Vec3 {
         self.a + self.b * t
     }
 }
